+++ contracts/userInterfaces/ftdc/ITeeKeyExistence.sol
@@ -38,7 +38,7 @@ interface ITeeKeyExistence {
         uint16 thresholdBIPS;
         uint64 timestamp;
         address[] cosigners;
-        uint256 cosignersThreshold;
+        uint64 cosignersThreshold;
         RequestBody requestBody;
         ResponseBody responseBody;
     }
